Streaming: integrate streaming into botFrameworkAdapter though useWebSocket and useNamedPipe methods
